feat:refatoração de exame
